[actions] Squash integer overflow bugs in alpha blending actions
There were some serious integer overflow bugs present in these actions,
due to our rampant conversion of alpha values to and fro signed and
unsigned numerical types (int16, uint8, real32). This commit resolves
the issue by converting the output (floating-point) to a value that is
always within range of an unsigned 8-bit integer, thus allowing us to
stay with floating-point numbers internally within the actions and
doing the conversion one-time only.

It's worth noting that I believe some of the tests were testing for the
wrong alpha output values -- reversed from what it should have been
testing for! Not sure how this happened, but alas! I feel confident
that the tests are proper now.

+++ tests/src/actions/ActionTest.cpp
@@ -279,7 +279,7 @@ void ActionTest::expected_action_params(  const SequenceAction* action,
 
 void
 ActionTest::expected_alpha_in_params( const FadeInAction* obj,
-                                      int16 alpha, const Sprite* tex,
+                                      uint8 alpha, const Sprite* tex,
                                       const std::string& scope_name )
 {
   ASSERT_TRUE(obj != nullptr)
@@ -298,7 +298,7 @@ ActionTest::expected_alpha_in_params( const FadeInAction* obj,
 
 void
 ActionTest::expected_alpha_out_params(  const FadeOutAction* obj,
-                                        int16 alpha, const Sprite* tex,
+                                        uint8 alpha, const Sprite* tex,
                                         const std::string& scope_name )
 {
   ASSERT_TRUE(obj != nullptr)
@@ -317,7 +317,7 @@ ActionTest::expected_alpha_out_params(  const FadeOutAction* obj,
 
 void
 ActionTest::expected_alpha_by_params( const FadeAlphaByAction* obj,
-                                      int16 alpha, const Sprite* tex,
+                                      uint8 alpha, const Sprite* tex,
                                       const std::string& scope_name )
 {
   ASSERT_TRUE(obj != nullptr)
